webstack/backend/requirements.txt
appdirs==1.4.4
autoflake==1.4
black==20.8b1
click==7.1.2
dbus-python==1.2.16
fastapi==0.62.0
flake8==3.8.4
h11==0.11.0
importlib-metadata==3.3.0
isort==5.6.4
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pathspec==0.8.1
pkg-resources==0.0.0
psutil==5.7.3
pycodestyle==2.6.0
pydantic==1.7.3
pyflakes==2.2.0
python-dotenv==0.15.0
regex==2020.11.13
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.13.1
zipp==3.4.0
